+++ cash/src/main/java/com/bank/cash/dtos/CashRequest.java
@@ -1,9 +1,11 @@
 package com.bank.cash.dtos;
 
+import java.math.BigDecimal;
+
 public class CashRequest {
     private String login;
 
-    private double delta;
+    private BigDecimal delta;
 
     public String getLogin() {
         return login;
@@ -13,11 +15,11 @@ public class CashRequest {
         this.login = login;
     }
 
-    public double getDelta() {
+    public BigDecimal getDelta() {
         return delta;
     }
 
-    public void setDelta(double delta) {
+    public void setDelta(BigDecimal delta) {
         this.delta = delta;
     }
 }
